+++ lib/eagle_provider.dart
@@ -1,4 +1,8 @@
 library eagle_provider;
 
 export 'src/controller.dart';
+export 'src/controller.dart';
+export 'src/controller_provider.dart';
 export 'src/state_controller.dart';
+export 'src/widgets/controller_builder.dart';
+export 'src/widgets/controller_consummer.dart';