+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2.0.0
+ - [add]: Add property ```listener``` in ```ControllerConsummer```
+ - [change]: Change concepted property ```listenWhen``` in ```ControllerConsummer```
+ - [remove]: Removed property ```listenWhem``` of ```ControllerBuilder``` Widget
+ - [fix]: Existing bug fix in status update
 ## 1.1.1
  - [fix]: Fixed issue with builder When
 ## 1.1.0